update test for dual epoch map and move test spec to top of the crate
