+++ src/stream-model/src/lib.rs
@@ -17,3 +17,39 @@ pub mod k8 {
         pub use k8_obj_metadata::*;
     }
 }
+
+
+
+#[cfg(test)]
+pub(crate) mod test_fixture {
+
+
+    use crate::core::{Spec, Status};
+    use crate::store::{ DefaultMetadataObject};
+    use crate::epoch::DualEpochMap;
+
+    // define test spec and status
+    #[derive(Debug, Default, Clone, PartialEq)]
+    pub struct TestSpec {
+        pub replica: u16,
+    }
+
+    impl Spec for TestSpec {
+        const LABEL: &'static str = "Test";
+        type IndexKey = String;
+        type Owner = Self;
+        type Status = TestStatus;
+    }
+
+    #[derive(Debug, Default, Clone, PartialEq)]
+    pub struct TestStatus {
+        pub up: bool,
+    }
+
+    impl Status for TestStatus {}
+
+    pub type DefaultTest = DefaultMetadataObject<TestSpec>;
+
+    pub type TestEpochMap = DualEpochMap<String, DefaultTest>;
+
+}
